Looks like hub not possible.
Interrupts mean host needs a fixed EP address.
MIDI device also needs fixed EP address.
Those two EP addresses have to align.
Hub also has EP address, and will conflict
with device EP address. Polling the hub
using the current approach causes the
pipe to switch to the hub and breaks
events on the MIDI device pipe, so that
is not an option if the EP on the device
is the same as the EP on the hub.